add shadow to blinn phong, changed floats to doubles for more precision

+++ libs/Hitable.h
@@ -6,7 +6,7 @@
 
 //struct to the point that was hit
 struct hit_record {
-	float t; //the t at the point
+	double t; //the t at the point
   Point3  p; //the point that was hit
   Vector3 normal; //the normal of the point in relation to the object
 	Material *material; //the material that was hit
@@ -22,7 +22,7 @@ public:
   Point3 get_center(){return center;}
 	Material* get_material(){return material;}
 
-  virtual bool hit(const Ray & r, float t_min, float t_max, hit_record & rec) const = 0;
+  virtual bool hit(const Ray & r, double t_min, double t_max, hit_record & rec) const = 0;
 };
 
 #endif
